Created a manager class that allows people to easily add the healthcheck and scheduledexecutor without too much effort.  Renamed the Original MainMethod class to SampleApplication.  Also removed some unused code.
