--- src/main/java/me/atam/atam4j/configuration/AcceptanceTestsBuildConfiguration.java
@@ -1,22 +1,0 @@
-package me.atam.atam4j.configuration;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-/**
- * @author anuragkapur
- */
-public class AcceptanceTestsBuildConfiguration implements AcceptanceTestsConfiguration {
-
-    @JsonProperty
-    private String helloWorldMessage;
-
-    @Override
-    public String getHelloWorldMessage() {
-        return helloWorldMessage;
-    }
-
-    @Override
-    public void setHelloWorldMessage(String helloWorldMessage) {
-        this.helloWorldMessage = helloWorldMessage;
-    }
-}
