--- src/main/java/me/atam/atam4j/configuration/AcceptanceTestsConfiguration.java
@@ -1,14 +1,0 @@
-package me.atam.atam4j.configuration;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-/**
- * @author anuragkapur
- */
-@JsonIgnoreProperties(ignoreUnknown=true)
-public interface AcceptanceTestsConfiguration {
-
-    public String getHelloWorldMessage();
-
-    public void setHelloWorldMessage(String helloWorldMessage);
-}
