--- src/main/java/me/atam/atam4j/tests/RunAllTest.java
@@ -1,10 +1,0 @@
-package me.atam.atam4j.tests;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-import org.junit.runner.RunWith;
-
-@RunWith(Cucumber.class)
-@CucumberOptions(monochrome = true)
-public class RunAllTest {
-}
